Fix the build on Swift 6.3 and move CI to checkout v5
Xcode 26.6 (Swift 6.3.3) on the CI runner does not convert Double to CGFloat
inside a tuple, which the local Swift 6.4 accepts, so the drag start offset is
converted explicitly. actions/checkout v5 runs on Node 24 instead of the
deprecated Node 20.

+++ Sources/MYUStats/Edge/EdgeController.swift
@@ -150,7 +150,7 @@ final class EdgeController {
                 return true
             }
             if state.isHandleHovered {
-                dragStart = (NSEvent.mouseLocation, UserDefaults.standard.double(forKey: SettingsKey.verticalOffset))
+                dragStart = (NSEvent.mouseLocation, CGFloat(UserDefaults.standard.double(forKey: SettingsKey.verticalOffset)))
                 state.isDragging = true
                 cancelHide()
                 if state.hovered != nil { setHovered(nil) }
